refactor: use Vite native ?raw imports, remove rawSvgPlugin
- Add ?raw suffix to all 42 SVG imports in libs/react/icons
  (Vite handles ?raw natively—no plugin needed for dev/build)
- Delete rawSvgPlugin from libs/vite-utils (no longer needed)
- Simplify svg-test-resolver: single svgTestResolver export
- Update shelter-web, wildfires, betterangels-admin vite configs
  to use Vite native SVG handling + test-only stub via process.env.VITEST

The only remaining custom SVG handling is the test-only resolveId
plugin that prevents cross-package Denied ID errors during Vitest.
