+++ apps/shelter-web/vite.config.mts
@@ -3,7 +3,7 @@ import tailwindcss from '@tailwindcss/postcss';
 import react from '@vitejs/plugin-react';
 import path from 'path';
 import { ProxyOptions, defineConfig } from 'vite';
-import { rawSvgPlugin, monorepoTsconfigAliases, svgTestResolverIfVitest } from '../../libs/vite-utils/src/index';
+import { monorepoTsconfigAliases, svgTestResolver } from '../../libs/vite-utils/src/index';
 import { baseHrefPlugin, getBranchBasePath } from '../../tools/shared/get-base-path.mjs';
 
 const SERVER_PORT = 8083;
@@ -47,9 +47,10 @@ export default defineConfig(({ mode }) => {
 
     plugins: [
       react(),
-      rawSvgPlugin(),
       baseHrefPlugin(basePath),
-      ...svgTestResolverIfVitest(__dirname),
+      // Vite handles ?raw SVG imports natively.
+      // Only stub SVGs during Vitest runs (avoids cross-package Denied ID).
+      ...(process.env.VITEST ? [svgTestResolver(__dirname)] : []),
     ],
 
     resolve: {
