+++ apps/wildfires/vite.config.mts
@@ -3,7 +3,7 @@ import tailwindcss from '@tailwindcss/postcss';
 import react from '@vitejs/plugin-react';
 import path from 'path';
 import { defineConfig } from 'vite';
-import { rawSvgPlugin, monorepoTsconfigAliases, svgTestResolverIfVitest } from '../../libs/vite-utils/src/index';
+import { monorepoTsconfigAliases, svgTestResolver } from '../../libs/vite-utils/src/index';
 import { baseHrefPlugin, getBranchBasePath } from '../../tools/shared/get-base-path.mjs';
 
 const SERVER_PORT = 8200;
@@ -33,9 +33,10 @@ export default defineConfig(({ mode }) => {
 
   plugins: [
     react(),
-    rawSvgPlugin(),
     baseHrefPlugin(basePath),
-    ...svgTestResolverIfVitest(__dirname),
+    // Vite handles ?raw SVG imports natively.
+    // Only stub SVGs during Vitest runs (avoids cross-package Denied ID).
+    ...(process.env.VITEST ? [svgTestResolver(__dirname)] : []),
   ],
 
   resolve: {
