+++ libs/react/icons/src/lib/components/betterAngelsLogoIcon.tsx
@@ -1,4 +1,4 @@
-import svg from '../../../../../assets/src/icons/svg/better_angels/betterAngelsLogo.svg';
+import svg from '../../../../../assets/src/icons/svg/better_angels/betterAngelsLogo.svg?raw';
 import { createSvgComponent } from '../../toComponent';
 
 export default createSvgComponent(svg);
