+++ libs/react/icons/src/lib/components/callOutlinedIcon.tsx
@@ -1,4 +1,4 @@
-import svg from '../../../../../../libs/expo/shared/icons/src/assets/third_party/mono_icons/svg/mi:call.svg';
+import svg from '../../../../../../libs/expo/shared/icons/src/assets/third_party/mono_icons/svg/mi:call.svg?raw';
 import { createSvgComponent } from '../../toComponent';
 
 export default createSvgComponent(svg);
