+++ libs/react/icons/src/lib/components/callRegularIcon.tsx
@@ -1,4 +1,4 @@
-import svg from '../../../../../../libs/expo/shared/icons/src/assets/third_party/microsoft_corporation/svg/fluent:call-20-regular.svg';
+import svg from '../../../../../../libs/expo/shared/icons/src/assets/third_party/microsoft_corporation/svg/fluent:call-20-regular.svg?raw';
 import { createSvgComponent } from '../../toComponent';
 
 export default createSvgComponent(svg);
