+++ libs/react/icons/src/lib/components/cameraIcon.tsx
@@ -1,4 +1,4 @@
-import svg from '../../../../../../libs/expo/shared/icons/src/assets/third_party/phosphor/svg/ph:camera.svg';
+import svg from '../../../../../../libs/expo/shared/icons/src/assets/third_party/phosphor/svg/ph:camera.svg?raw';
 import { createSvgComponent } from '../../toComponent';
 
 export default createSvgComponent(svg);
