+++ libs/react/icons/src/lib/components/checkIcon.tsx
@@ -1,4 +1,4 @@
-import svg from '../../../../../../libs/expo/shared/icons/src/assets/third_party/dariush_habibpour/svg/iconamoon:check-bold.svg';
+import svg from '../../../../../../libs/expo/shared/icons/src/assets/third_party/dariush_habibpour/svg/iconamoon:check-bold.svg?raw';
 import { createSvgComponent } from '../../toComponent';
 
 export default createSvgComponent(svg);
