+++ libs/react/icons/src/lib/components/chevronLeftIcon.tsx
@@ -1,4 +1,4 @@
-import svg from '../../../../../../libs/expo/shared/icons/src/assets/third_party/dave_gandy/svg/fa6-solid:chevron-left.svg';
+import svg from '../../../../../../libs/expo/shared/icons/src/assets/third_party/dave_gandy/svg/fa6-solid:chevron-left.svg?raw';
 
 import { createSvgComponent } from '../../toComponent';
