+++ libs/react/icons/src/lib/components/clockIcon.tsx
@@ -1,4 +1,4 @@
-import svg from '../../../../../../libs/expo/shared/icons/src/assets/third_party/mingcute_design/svg/mingcute:time-line.svg';
+import svg from '../../../../../../libs/expo/shared/icons/src/assets/third_party/mingcute_design/svg/mingcute:time-line.svg?raw';
 import { createSvgComponent } from '../../toComponent';
 
 export default createSvgComponent(svg);
