+++ libs/react/icons/src/lib/components/closeIcon.tsx
@@ -1,4 +1,4 @@
-import svg from '../../../../../../libs/expo/shared/icons/src/assets/third_party/mingcute_design/svg/mingcute:close-fill.svg';
+import svg from '../../../../../../libs/expo/shared/icons/src/assets/third_party/mingcute_design/svg/mingcute:close-fill.svg?raw';
 import { createSvgComponent } from '../../toComponent';
 
 export default createSvgComponent(svg);
