+++ libs/react/icons/src/lib/components/deoIcon.tsx
@@ -1,4 +1,4 @@
-import svg from '../../../../../../libs/expo/shared/icons/src/assets/third_party/wildfire_link_logos/deo-logo-color.svg';
+import svg from '../../../../../../libs/expo/shared/icons/src/assets/third_party/wildfire_link_logos/deo-logo-color.svg?raw';
 import { createSvgComponent } from '../../toComponent';
 
 export default createSvgComponent(svg);
