+++ libs/react/icons/src/lib/components/eddIcon.tsx
@@ -1,4 +1,4 @@
-import svg from '../../../../../../libs/expo/shared/icons/src/assets/third_party/wildfire_link_logos/EDD-logo.svg';
+import svg from '../../../../../../libs/expo/shared/icons/src/assets/third_party/wildfire_link_logos/EDD-logo.svg?raw';
 import { createSvgComponent } from '../../toComponent';
 
 export default createSvgComponent(svg);
