+++ libs/react/icons/src/lib/components/emailIcon.tsx
@@ -1,4 +1,4 @@
-import svg from '../../../../../../libs/expo/shared/icons/src/assets/third_party/material_design_authors/svg/ic:outline-email.svg';
+import svg from '../../../../../../libs/expo/shared/icons/src/assets/third_party/material_design_authors/svg/ic:outline-email.svg?raw';
 import { createSvgComponent } from '../../toComponent';
 
 export default createSvgComponent(svg);
