+++ libs/react/icons/src/lib/components/emailOutlinedIcon.tsx
@@ -1,4 +1,4 @@
-import svg from '../../../../../../libs/expo/shared/icons/src/assets/third_party/we_ui/weui:email-outlined.svg';
+import svg from '../../../../../../libs/expo/shared/icons/src/assets/third_party/we_ui/weui:email-outlined.svg?raw';
 import { createSvgComponent } from '../../toComponent';
 
 export default createSvgComponent(svg);
