+++ libs/react/icons/src/lib/components/faceFrownIcon.tsx
@@ -1,4 +1,4 @@
-import svg from '../../../../../../libs/expo/shared/icons/src/assets/third_party/dave_gandy/svg/fa6-regular:face-frown.svg';
+import svg from '../../../../../../libs/expo/shared/icons/src/assets/third_party/dave_gandy/svg/fa6-regular:face-frown.svg?raw';
 
 import { createSvgComponent } from '../../toComponent';
