+++ libs/react/icons/src/lib/components/filterIcon.tsx
@@ -1,4 +1,4 @@
-import svg from '../../../../../assets/src/icons/svg/filterOutline.svg';
+import svg from '../../../../../assets/src/icons/svg/filterOutline.svg?raw';
 import { createSvgComponent } from '../../toComponent';
 
 export default createSvgComponent(svg);
