+++ libs/react/icons/src/lib/components/globeIcon.tsx
@@ -1,4 +1,4 @@
-import svg from '../../../../../../libs/expo/shared/icons/src/assets/third_party/arturo_wibawa/svg/akar-icons:globe.svg';
+import svg from '../../../../../../libs/expo/shared/icons/src/assets/third_party/arturo_wibawa/svg/akar-icons:globe.svg?raw';
 import { createSvgComponent } from '../../toComponent';
 
 export default createSvgComponent(svg);
