+++ libs/react/icons/src/lib/components/groupsIcon.tsx
@@ -1,4 +1,4 @@
-import svg from '../../../../../../libs/expo/shared/icons/src/assets/third_party/google/svg/material-symbols:groups-outline.svg';
+import svg from '../../../../../../libs/expo/shared/icons/src/assets/third_party/google/svg/material-symbols:groups-outline.svg?raw';
 import { createSvgComponent } from '../../toComponent';
 
 export default createSvgComponent(svg);
