+++ libs/react/icons/src/lib/components/imageIcon.tsx
@@ -1,4 +1,4 @@
-import svg from '../../../../../../libs/expo/shared/icons/src/assets/third_party/hugeicons/hugeicons:city-03.svg';
+import svg from '../../../../../../libs/expo/shared/icons/src/assets/third_party/hugeicons/hugeicons:city-03.svg?raw';
 import { createSvgComponent } from '../../toComponent';
 
 export default createSvgComponent(svg);
