+++ libs/react/icons/src/lib/components/instagramIcon.tsx
@@ -1,4 +1,4 @@
-import svg from '../../../../../../libs/expo/shared/icons/src/assets/third_party/remix_design/svg/ri:instagram-line.svg';
+import svg from '../../../../../../libs/expo/shared/icons/src/assets/third_party/remix_design/svg/ri:instagram-line.svg?raw';
 import { createSvgComponent } from '../../toComponent';
 
 export default createSvgComponent(svg);
