+++ libs/react/icons/src/lib/components/lightBulbIcon.tsx
@@ -1,4 +1,4 @@
-import svg from '../../../../../../libs/expo/shared/icons/src/assets/third_party/hugeicons/hugeicons:idea-01.svg';
+import svg from '../../../../../../libs/expo/shared/icons/src/assets/third_party/hugeicons/hugeicons:idea-01.svg?raw';
 
 import { createSvgComponent } from '../../toComponent';
