+++ libs/react/icons/src/lib/components/listIcon.tsx
@@ -1,4 +1,4 @@
-import svg from '../../../../../../libs/expo/shared/icons/src/assets/third_party/pictogrammers/svg/mdi:format-list-bulleted.svg';
+import svg from '../../../../../../libs/expo/shared/icons/src/assets/third_party/pictogrammers/svg/mdi:format-list-bulleted.svg?raw';
 import { createSvgComponent } from '../../toComponent';
 
 export default createSvgComponent(svg);
