+++ libs/react/icons/src/lib/components/locationIcon.tsx
@@ -1,4 +1,4 @@
-import svg from '../../../../../assets/src/icons/svg/locationIcon.svg';
+import svg from '../../../../../assets/src/icons/svg/locationIcon.svg?raw';
 import { createSvgComponent } from '../../toComponent';
 
 export default createSvgComponent(svg);
