+++ libs/react/icons/src/lib/components/menuIcon.tsx
@@ -1,4 +1,4 @@
-import svg from '../../../../../../libs/expo/shared/icons/src/assets/third_party/dave_gandy/svg/fa6-solid:bars.svg';
+import svg from '../../../../../../libs/expo/shared/icons/src/assets/third_party/dave_gandy/svg/fa6-solid:bars.svg?raw';
 
 import { createSvgComponent } from '../../toComponent';
