+++ libs/react/icons/src/lib/components/plusIcon.tsx
@@ -1,4 +1,4 @@
-import svg from '../../../../../../libs/expo/shared/icons/src/assets/third_party/dave_gandy/svg/fa6-solid:plus.svg';
+import svg from '../../../../../../libs/expo/shared/icons/src/assets/third_party/dave_gandy/svg/fa6-solid:plus.svg?raw';
 
 import { createSvgComponent } from '../../toComponent';
