+++ libs/react/icons/src/lib/components/shareIcon.tsx
@@ -1,4 +1,4 @@
-import svg from '../../../../../assets/src/icons/svg/share.svg';
+import svg from '../../../../../assets/src/icons/svg/share.svg?raw';
 import { createSvgComponent } from '../../toComponent';
 
 export default createSvgComponent(svg);
