+++ libs/react/icons/src/lib/components/threeDotIcon.tsx
@@ -1,4 +1,4 @@
-import svg from '../../../../../../libs/expo/shared/icons/src/assets/third_party/pictogrammers/svg/mdi:ellipsis-vertical.svg';
+import svg from '../../../../../../libs/expo/shared/icons/src/assets/third_party/pictogrammers/svg/mdi:ellipsis-vertical.svg?raw';
 import { createSvgComponent } from '../../toComponent';
 
 export default createSvgComponent(svg);
