+++ libs/react/icons/src/lib/components/userIcon.tsx
@@ -1,4 +1,4 @@
-import svg from '../../../../../../libs/expo/shared/icons/src/assets/third_party/pictogrammers/svg/mdi:user-outline.svg';
+import svg from '../../../../../../libs/expo/shared/icons/src/assets/third_party/pictogrammers/svg/mdi:user-outline.svg?raw';
 
 import { createSvgComponent } from '../../toComponent';
