+++ libs/react/icons/src/lib/components/usersIcon.tsx
@@ -1,4 +1,4 @@
-import svg from '../../../../../../libs/expo/shared/icons/src/assets/third_party/pictogrammers/svg/mdi:users-outline.svg';
+import svg from '../../../../../../libs/expo/shared/icons/src/assets/third_party/pictogrammers/svg/mdi:users-outline.svg?raw';
 
 import { createSvgComponent } from '../../toComponent';
