+++ libs/react/icons/src/lib/components/wfAlertOn.tsx
@@ -1,4 +1,4 @@
-import svg from '../../../../../../libs/expo/shared/icons/src/assets/third_party/microsoft_corporation/svg/fluent:alert-on-regular.svg';
+import svg from '../../../../../../libs/expo/shared/icons/src/assets/third_party/microsoft_corporation/svg/fluent:alert-on-regular.svg?raw';
 
 import { createSvgComponent } from '../../toComponent';
