+++ libs/react/icons/src/lib/components/wfAnnouncement.tsx
@@ -1,4 +1,4 @@
-import svg from '../../../../../../libs/expo/shared/icons/src/assets/third_party/pictogrammers/svg/mdi:announcement.svg';
+import svg from '../../../../../../libs/expo/shared/icons/src/assets/third_party/pictogrammers/svg/mdi:announcement.svg?raw';
 
 import { createSvgComponent } from '../../toComponent';
