+++ libs/react/icons/src/lib/components/wfLinkIcon.tsx
@@ -1,4 +1,4 @@
-import svg from '../../../../../../libs/expo/shared/icons/src/assets/third_party/hugeicons/hugeicons:link-square-02.svg';
+import svg from '../../../../../../libs/expo/shared/icons/src/assets/third_party/hugeicons/hugeicons:link-square-02.svg?raw';
 
 import { createSvgComponent } from '../../toComponent';
